docs(button): refactor button demo
re #413
