--- src/button/demos/shape.vue
@@ -1,8 +1,0 @@
-<template>
-  <div>
-    <t-button>长方形按钮</t-button>
-    <t-button shape="square">正方形按钮</t-button>
-    <t-button shape="round">圆角长方形按钮</t-button>
-    <t-button shape="circle">圆形按钮</t-button>
-  </div>
-</template>
